paired tick E4 + overlapped restart: copy-engine prefetch into a device ring (host pack per (request, block) both heads, one cudaMemcpyAsync per layer on a side stream, event readiness, D2D ring-hit kernel between diff and the gathers, full accounting), NOSI_PAIRED_PREFETCH=off|ring; the two-position catch-up on a side stream with its own captured scoring graph, NOSI_PAIRED_RESTART=sequential|overlapped; twin2b forces gemm=fused; the in-situ twin casts the fused prefix
prefetch.py: plan_prefetch (want = sel_S minus resident minus in-flight minus ids >= T, unique (b, blk) over
heads, canonical, cap), pack_rows / pack_pieces (index_select of contiguous 32 KB runs), id_table, RingBook
(double-buffered halves per layer: issue / late / arrive / consume / full / release), diff_reference (the CUDA
kernel's rule), ring_hit (Triton, grid (H, B, M): D2D scatter of the served pieces + bias rows from total_cis,
clears _load_mask, counts) + reference, v_update_with_ring (decode_update_has_kv_bias replayed with the ring
step where the pool sits), PrefetchAccount, PrefetchEngine (guarded: worker pool, pinned staging, side stream).
core.py: CatchupState (ACTIVE / CATCHING, tok1 judged at the next tick's end, rejoin at t + 2), catchup_schedule.
side_restart.py: SideScoring (second capture of the pooling / top-k graph over private buffers per layer),
CatchupRunner (position 1 interleaved per layer behind the main tick's layer events; position 2 after its head;
the next tick waits per layer on the side events). tick.py: ScoringSet, tick_index, before/after-layer hooks,
tv / own-row overrides, setup_prefetch, the ring step in the V update. Pilot: knobs, prefetch summary,
catch-up timing, main-stream syncs when a side stream runs. DESIGN.md sections 9 / 10 with registered
predictions. Tests 70 (8 new) incl. the ring-hit kernel through the interpreter and the ring-vs-plain body.

+++ nosi/nosi/paired/twin.py
@@ -121,7 +121,7 @@ class InSituTwin:
         U = q.shape[1]
         # term: bias / tail rule
         cbi = torch.arange(B, dtype=torch.int32, device=x.device)
-        rb1 = core.paired_bias(eng._kv_bias_gpu, vis_v, cbi, 1, cfg.masked, out=sc.bias_twin, check_values=False)
+        rb1 = core.paired_bias(eng._kv_bias_gpu, vis_v.to(torch.int64), cbi, 1, cfg.masked, out=sc.bias_twin, check_values=False)   # the fused path's prefix is int32 (job 2175574)
         rows_v = torch.arange(0, U * B, U, device=x.device)
         rec["bias_equal"] = bool(torch.equal(rb.bias[rows_v], rb1.bias[:B]) and torch.equal(rb.cache_seqlens[rows_v], rb1.cache_seqlens)
                                  and torch.equal(rb.cache_batch_idx[rows_v], rb1.cache_batch_idx))
